+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - [docker](src/docker/bin/entrypoint.sh) Now use entry.d and exit.d.
 - [docker](src/docker/bin/docker-common.sh) Fixed minor issue in logging functionality.
+- [demo](demo) Made service names shorter.
 
 # 1.3.7